New field (targetAddress) in Route message, now the destAddress in a broadcast message is the broadcastAddress and the target should be in the payload.

+++ nRFTP/IPhysicalLayer.h
@@ -11,7 +11,7 @@ namespace nRFTP{
 
   class IPhysicalLayer{
     public:
-      virtual bool write( const void* buf, uint8_t len, uint64_t destAddress) = 0;
+      virtual bool write( const void* buf, uint8_t len, uint16_t destAddress) = 0;
       virtual bool available(void) = 0;
       virtual bool read( void* buf, uint8_t len ) = 0;
       virtual void begin(void) = 0;
